requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
joblib==1.0.1
jsonpointer==2.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.1
torchfile==0.1.0
torchvision==0.9.1
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==1.0.1

